Minor pairing changes
